requirements.txt
dash==2.18.2
dash_iconify==0.1.2
dash_mantine_components==0.15.0
flask_caching==2.3.0
numpy==2.1.3
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
mysqlclient==2.2.6
dash-new-tab==0.0.3